+++ python/cugraph/__init__.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2019-2020, NVIDIA CORPORATION.
+# Copyright (c) 2019-2021, NVIDIA CORPORATION.
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
@@ -24,11 +24,15 @@ from cugraph.community import (
     analyzeClustering_ratio_cut,
     subgraph,
     triangles,
+    ego_graph,
+    batched_ego_graphs,
 )
 
 from cugraph.structure import (
     Graph,
     DiGraph,
+    MultiGraph,
+    MultiDiGraph,
     from_edgelist,
     from_cudf_edgelist,
     from_pandas_edgelist,
@@ -77,7 +81,8 @@ from cugraph.traversal import (
     sssp,
     shortest_path,
     filter_unreachable,
-    shortest_path_length
+    shortest_path_length,
+    traveling_salesperson
 )
 
 from cugraph.tree import minimum_spanning_tree, maximum_spanning_tree
@@ -89,7 +94,7 @@ from cugraph.bsp.traversal import bfs_df_pregel
 from cugraph.proto.components import strong_connected_component
 from cugraph.proto.structure import find_bicliques
 
-from cugraph.linear_assignment import hungarian
+from cugraph.linear_assignment import hungarian, dense_hungarian
 from cugraph.layout import force_atlas2
 from cugraph.raft import raft_include_test
 from cugraph.comms import comms