+++ python/cugraph/link_analysis/pagerank_wrapper.pyx
@@ -1,4 +1,4 @@
-# Copyright (c) 2019-2020, NVIDIA CORPORATION.
+# Copyright (c) 2019-2021, NVIDIA CORPORATION.
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
@@ -33,21 +33,22 @@ def pagerank(input_graph, alpha=0.85, personalization=None, max_iter=100, tol=1.
     Call pagerank
     """
 
-    if not input_graph.transposedadjlist:
-        input_graph.view_transposed_adj_list()
-
     cdef unique_ptr[handle_t] handle_ptr
     handle_ptr.reset(new handle_t())
     handle_ = handle_ptr.get();
 
-    [offsets, indices] = graph_primtypes_wrapper.datatype_cast([input_graph.transposedadjlist.offsets, input_graph.transposedadjlist.indices], [np.int32])
-    [weights] = graph_primtypes_wrapper.datatype_cast([input_graph.transposedadjlist.weights], [np.float32, np.float64])
+    [src, dst] = graph_primtypes_wrapper.datatype_cast([input_graph.edgelist.edgelist_df['src'], input_graph.edgelist.edgelist_df['dst']], [np.int32])
+    weights = None
+    if input_graph.edgelist.weights:
+        [weights] = graph_primtypes_wrapper.datatype_cast([input_graph.edgelist.edgelist_df['weights']], [np.float32, np.float64])
 
     num_verts = input_graph.number_of_vertices()
     num_edges = input_graph.number_of_edges(directed_edges=True)
+    # FIXME: needs to be edge_t type not int
+    cdef int num_partition_edges = len(src)
 
     df = cudf.DataFrame()
-    df['vertex'] = cudf.Series(np.zeros(num_verts, dtype=np.int32))
+    df['vertex'] = cudf.Series(np.arange(num_verts, dtype=np.int32))
     df['pagerank'] = cudf.Series(np.zeros(num_verts, dtype=np.float32))
 
     cdef bool has_guess = <bool> 0
@@ -62,19 +63,16 @@ def pagerank(input_graph, alpha=0.85, personalization=None, max_iter=100, tol=1.
 
     cdef uintptr_t c_pers_vtx = <uintptr_t>NULL
     cdef uintptr_t c_pers_val = <uintptr_t>NULL
-    cdef sz = 0
+    cdef int sz = 0
 
-    cdef uintptr_t c_offsets = offsets.__cuda_array_interface__['data'][0]
-    cdef uintptr_t c_indices = indices.__cuda_array_interface__['data'][0]
-    cdef uintptr_t c_weights = <uintptr_t>NULL
-    cdef uintptr_t c_local_verts = <uintptr_t> NULL;
-    cdef uintptr_t c_local_edges = <uintptr_t> NULL;
-    cdef uintptr_t c_local_offsets = <uintptr_t> NULL;
-
+    cdef uintptr_t c_src_vertices = src.__cuda_array_interface__['data'][0]
+    cdef uintptr_t c_dst_vertices = dst.__cuda_array_interface__['data'][0]
+    cdef uintptr_t c_edge_weights = <uintptr_t>NULL
+    
     personalization_id_series = None
 
     if weights is not None:
-        c_weights = weights.__cuda_array_interface__['data'][0]
+        c_edge_weights = weights.__cuda_array_interface__['data'][0]
         weight_t = weights.dtype
     else:
         weight_t = np.dtype("float32")
@@ -94,15 +92,19 @@ def pagerank(input_graph, alpha=0.85, personalization=None, max_iter=100, tol=1.
         c_pers_val = personalization['values'].__cuda_array_interface__['data'][0]
 
     cdef graph_container_t graph_container
-    populate_graph_container_legacy(graph_container,
-                                    <graphTypeEnum>(<int>(graphTypeEnum.LegacyCSC)),
-                                    handle_[0],
-                                    <void*>c_offsets, <void*>c_indices, <void*>c_weights,
-                                    <numberTypeEnum>(<int>(numberTypeEnum.int32Type)),
-                                    <numberTypeEnum>(<int>(numberTypeEnum.int32Type)),
-                                    <numberTypeEnum>(<int>(numberTypeMap[weight_t])),
-                                    num_verts, num_edges,
-                                    <int*>c_local_verts, <int*>c_local_edges, <int*>c_local_offsets)
+    populate_graph_container(graph_container,
+                             handle_[0],
+                             <void*>c_src_vertices, <void*>c_dst_vertices, <void*>c_edge_weights,
+                             <void*>NULL,
+                             <numberTypeEnum>(<int>(numberTypeEnum.int32Type)),
+                             <numberTypeEnum>(<int>(numberTypeEnum.int32Type)),
+                             <numberTypeEnum>(<int>(numberTypeMap[weight_t])),
+                             #num_verts, num_edges,
+                             num_partition_edges,
+                             num_verts, num_edges,
+                             False,
+                             True,
+                             False)
 
     if (df['pagerank'].dtype == np.float32):
         call_pagerank[int, float](handle_[0], graph_container,